+++ src/framework/piral-core/src/actions/app.ts
@@ -36,6 +36,11 @@ export function injectPilet(ctx: GlobalStateContext, pilet: Pilet) {
     modules: replaceOrAddItem(state.modules, pilet, m => m.name === pilet.name),
     registry: removeNested<RegistryState, BaseRegistration>(state.registry, m => m.pilet === pilet.name),
   }));
+
+  ctx.emit('unload-pilet', {
+    name: pilet.name,
+    version: pilet.version,
+  });
 }
 
 export function setComponent<TKey extends keyof ComponentsState>(